refactor: move modelo StudentGift do app api para students
- O modelo StudentGift e sua migration foram removidos do app api e adicionados ao app students para melhor organização
- Todos os imports que referenciavam StudentGift foram atualizados
